feat(searchbar)!: remove top level subcomponents
