+++ src/components/SearchBar/SearchBar.stories.tsx
@@ -8,7 +8,7 @@ export default {
   title: 'Components/SearchBar',
   component: SearchBar,
   parameters: {
-    badges: ['1.1', BADGE.BETA],
+    badges: ['1.1', BADGE.NEEDS_REVISION],
   },
   decorators: [
     (Story) => (
