--- src/components/SearchButton/SearchButton.module.css
@@ -1,14 +1,0 @@
-/*------------------------------------*\
-    # SEARCH BUTTON
-\*------------------------------------*/
-
-/**
- * Button to trigger search in a Search Bar.
- */
-.search-button {
-  padding-top: 1px;
-  padding-bottom: 1px;
-  height: var(--eds-size-5);
-  box-sizing: content-box;
-  border: 0;
-}
