--- src/components/SearchButton/index.ts
@@ -1,1 +1,0 @@
-export { SearchButton as default } from './SearchButton';
