--- src/components/SearchField/SearchField.module.css
@@ -1,45 +1,0 @@
-/*------------------------------------*\
-    # SEARCH FIELD
-\*------------------------------------*/
-
-/**
- * Search Input Field for gathering input text.
- */
-.search-field {
-  flex: 1;
-  position: relative;
-}
-
-/**
- * The Input component.
- */
-.search-field > .search-field__input {
-  padding-left: 2.25rem;
-}
-.search-field__input::-webkit-search-cancel-button {
-  display: none;
-}
-
-/**
- * The search icon that overlays the Input.
- */
-.search-field__icon {
-  color: var(--eds-theme-color-icon-neutral-default);
-  position: absolute;
-  top: 0.6875rem;
-  left: 0.625rem;
-}
-
-/**
- * Focused variant of the Search Icon.
- */
-.search-field__input:focus + .search-field__icon {
-  color: var(--eds-theme-color-icon-brand-primary);
-}
-
-/**
- * Disabled variant of the Search Icon.
- */
-.search-field__icon--disabled {
-  color: var(--eds-theme-color-icon-disabled);
-}
