--- src/components/SearchField/index.ts
@@ -1,1 +1,0 @@
-export { SearchField as default } from './SearchField';
